requirements.txt
google-generativeai
python-dotenv
pillow
pandas
requests
